backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
httpx
pypdf
pymongo
dnspython
gunicorn
